// client/src/enums/ButtonEnum.ts
export enum ButtonClassType {
  PRIMARY = 'button',
  SECONDARY = 'button__secondary',
}

export enum ButtonClassModifier {
  NEXT_PAGE = 'button__secondary--next-page',
  PREV_PAGE = 'button__secondary--prev-page',
  UP_PAGE = 'button__secondary--up-page',
  NUM_PAGE = 'button__secondary--num-page',
  HEART = 'button__secondary--heart',
  ADDED = 'button-added',
}

export enum ButtonType {
  CART = 'cart',
  HEART = 'heart',
}